requirements.txt
pyqrcode
astroid==2.3.3
bson==0.5.8
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
dataclasses==0.6
<<<<<<< HEAD
Django==2.2.13
django-rest-framework==0.1.0
=======
sqlparse==0.2.4
Django==2.2
django-multiforloop==0.2.1
django-phone-field==1.8.0
>>>>>>> origin/paras
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djongo==1.2.38
et-xmlfile==1.0.1
idna==2.9
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
lml==0.0.9
mccabe==0.6.1
passlib==1.7.2
Pillow==7.0.0
PyJWT==1.7.1
openpyxl==3.0.3
pyexcel-io==0.5.20
pyexcel-xlsx==0.5.8
pylint==2.4.4
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2